feat: content creator dashboard components
- refactor carousel component to a reusable component
- remove dashboard in components name
